training_pipeline/requirements.txt
numpy==1.26.4
google-cloud-storage
google-cloud-aiplatform
pandas
torch
diffusers
transformers
accelerate
safetensors
opencv-python